// src/api/generated/schemas/productDetail.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from './category';
import type { Tag } from './tag';
import type { StatusEnum } from './statusEnum';
import type { ProductImage } from './productImage';

/**
 * Detailed serializer for product detail views.
 */
export interface ProductDetail {
  readonly id: number;
  /**
   * Product name
   * @maxLength 200
   */
  name: string;
  /**
   * URL-friendly version of the name
   * @maxLength 200
   * @pattern ^[-a-zA-Z0-9_]+$
   */
  slug: string;
  /** Detailed product description */
  description: string;
  /**
   * Short description for listings
   * @maxLength 500
   */
  short_description?: string;
  /**
   * Product price
   * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
   */
  price: string;
  /**
   * Original price for showing discounts
   * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
   */
  original_price: string;
  readonly current_price: number;
  /** Calculate discount percentage. */
  readonly discount_percentage: number;
  /**
   * Unique product identifier
   * @maxLength 100
   */
  sku: string;
  /**
   * Available quantity in stock
   * @minimum 0
   * @maximum 2147483647
   */
  stock_quantity?: number;
  readonly category: Category;
  category_id: number;
  readonly tags: readonly Tag[];
  tag_ids?: number[];
  /** Product status

* `draft` - Draft
* `active` - Active
* `inactive` - Inactive
* `out_of_stock` - Out of Stock */
  status?: StatusEnum;
  /** Is product visible to the users */
  is_visible?: boolean;
  /** @nullable */
  sale_start?: string | null;
  /** @nullable */
  sale_end?: string | null;
  readonly is_on_sale: boolean;
  /** Check if product is in stock. */
  readonly is_in_stock: boolean;
  /** Check if product is available for purchase. */
  readonly is_available: boolean;
  readonly images: readonly ProductImage[];
  /** Timestamp when the record was created */
  readonly created_at: string;
  /** Timestamp when the record was last updated */
  readonly updated_at: string;
}